requirements.txt
Flask
tensorflow
numpy
matplotlib
scikit-learn
Pillow
Werkzeug
gunicorn
ultralytics